# frontend/scripts/openapi/contracts/auth.json
{
  "components": {
    "schemas": {
...
      "Credential": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "credentialId": {
            "format": "uuid",
            "title": "Credentialid",
            "type": "string"
          },
          "credentialType": {
            "$ref": "#/components/schemas/CredentialType"
          },
          "currentVersion": {
            "minimum": 1.0,
            "title": "Currentversion",
            "type": "integer"
          },
          "deletedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deletedat"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "lockVersion": {
            "default": 1,
            "minimum": 1.0,
            "title": "Lockversion",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "publicConfig": {
            "additionalProperties": true,
            "title": "Publicconfig",
            "type": "object"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "credentialId",
          "name",
          "credentialType",
          "username",
          "currentVersion",
          "enabled",
          "description",
          "createdAt",
          "updatedAt"
        ],
        "title": "Credential",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpsComposer API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/credentials": {
      "get": {
        "operationId": "listCredentials",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Credential"
                  },
                  "title": "Response Listcredentials",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Credentials",
        "tags": [
          "inventory"
        ]
      },
      "post": {
        "operationId": "createCredential",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "PASSWORD": "#/components/schemas/CredentialCreateRequest",
                    "SSH_PRIVATE_KEY": "#/components/schemas/SshPrivateKeyCredentialCreateRequest"
                  },
                  "propertyName": "credentialType"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CredentialCreateRequest"
                  },
                  {
                    "$ref": "#/components/schemas/SshPrivateKeyCredentialCreateRequest"
                  }
                ],
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Credential"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Credential",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/credentials/{credential_id}": {
      "delete": {
        "operationId": "deleteCredential",
        "parameters": [
          {
            "in": "path",
            "name": "credential_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Credential Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Credential",
        "tags": [
          "inventory"
        ]
      },
      "get": {
        "operationId": "getCredential",
        "parameters": [
          {
            "in": "path",
            "name": "credential_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Credential Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Credential"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Credential",
        "tags": [
          "inventory"
        ]
      },
      "put": {
        "operationId": "updateCredential",
        "parameters": [
          {
            "in": "path",
            "name": "credential_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Credential Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Credential"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Credential",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/credentials/{credential_id}/revisions": {
      "get": {
        "operationId": "listCredentialRevisions",
        "parameters": [
          {
            "in": "path",
            "name": "credential_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Credential Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CredentialRevisionResponse"
                  },
                  "title": "Response Listcredentialrevisions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Credential Revisions",
        "tags": [
          "inventory"
        ]
      },
      "post": {
        "operationId": "rotateCredential",
        "parameters": [
          {
            "in": "path",
            "name": "credential_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Credential Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "PASSWORD": "#/components/schemas/CredentialRotateRequest",
                    "SSH_PRIVATE_KEY": "#/components/schemas/SshPrivateKeyCredentialRotateRequest"
                  },
                  "propertyName": "credentialType"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CredentialRotateRequest"
                  },
                  {
                    "$ref": "#/components/schemas/SshPrivateKeyCredentialRotateRequest"
                  }
                ],
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Credential"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rotate Credential",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/groups": {
      "get": {
        "operationId": "listGroups",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HostGroup"
                  },
                  "title": "Response Listgroups",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Groups",
        "tags": [
          "inventory"
        ]
      },
      "post": {
        "operationId": "createGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostGroup"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Group",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/groups/{group_id}": {
      "delete": {
        "operationId": "deleteGroup",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Group Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Group",
        "tags": [
          "inventory"
        ]
      },
      "put": {
        "operationId": "updateGroup",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Group Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostGroup"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Group",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/hosts": {
      "get": {
        "operationId": "listHosts",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Host"
                  },
                  "title": "Response Listhosts",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Hosts",
        "tags": [
          "inventory"
        ]
      },
      "post": {
        "operationId": "createHost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Host"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Host",
        "tags": [
          "inventory"
        ]
      }
    },
...
    "/api/v1/hosts/{host_id}/host-keys": {
      "get": {
        "operationId": "listHostKeys",
        "parameters": [
          {
            "in": "path",
            "name": "host_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Host Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HostKey"
                  },
                  "title": "Response Listhostkeys",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Host Keys",
        "tags": [
          "inventory"
        ]
      }
    },
...
    "/api/v1/hosts/{host_id}/host-keys/scan": {
      "post": {
        "operationId": "scanHostKeys",
        "parameters": [
          {
            "in": "path",
            "name": "host_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Host Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HostKeyScanResponse"
                  },
                  "title": "Response Scanhostkeys",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scan Host Keys",
        "tags": [
          "inventory"
        ]
      }
    },
...
  }
}